Extract npm script for oneSimulation
