+++ README.md
@@ -24,6 +24,7 @@ npm run dev
 npm run logXpLevelInfo
 npm run logXpTable
 npm run getLevelFor10kXP
+npm run oneSimulation
 ```
 
 ## Formatting
